Rework LoRa code - combine gateway and node code together
